[FIX] Correct speed movement to match FPS
Because requestAnimateFrame is based on the refresh rate
of the current display the initial
values for player and enemies movements, originally set for 60Hz,
where ever too fast (144Hz screen) or too slow (if slow CPU not able
to maintain the fram rate).
Now we ajust the movement base on the actual FPS, using the 60Hz previous
default values as a base.